refactor: extract shadow convention plugin
