+++ gradle.properties
@@ -1,0 +1,1 @@
+systemProp.org.gradle.kotlin.dsl.precompiled.accessors.strict=true
